backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.9.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
passlib==1.7.4
bcrypt==4.1.2
